Updated docs based on modern experiences
I'm giving up on nVidia, but I wanted to get the documentation as close to correct as I could.  What was there was clearly not working. For example, yum has been replaced by dnf, the "xorg-x11-drv-nvidia" and "nvidia-kmod" packages don't exist in dnf, and so on.  I wish I could have gotten the documentation into better shape, doubly so since it'd mean that I wouldn't have to go buy a video card, but this is as far as I'm willing to go on this battle.  Best of luck to the next person trying to take a stab at nVidia and their bloody driver.